feat(no-issue):(add validation of empty variables) (#2)

+++ src/domain/usecase/account/AccountUseCase.java
@@ -2,6 +2,7 @@ package domain.usecase.account;
 
 import data.repository.account.AccountRepository;
 import domain.entities.account.AccountEntities;
+import shared.Utils;
 
 import java.util.ArrayList;
 
@@ -12,15 +13,32 @@ public class AccountUseCase {
         repository = new AccountRepository();
     }
 
+    private void verifyAllIsEmpty(String name, Double balance){
+        if (Utils.isEmptyParams(name)) {
+            throw new IllegalArgumentException("O [name] não pode ser vazio, digite corretamente");
+        }
+        if (Utils.isEmptyParams(balance.toString())) {
+            throw new IllegalArgumentException("O [balance] não pode ser vazio, digite corretamente");
+        }
+    }
+
     public AccountEntities create(String name, double balance) {
+        verifyAllIsEmpty(name, balance);
+
         return repository.createAccount(name, balance);
     }
 
     public AccountEntities edit(String name, double balance) {
+        verifyAllIsEmpty(name, balance);
+
         return repository.editAccount(name, balance);
     }
 
     public void remove(String name) {
+        if (Utils.isEmptyParams(name)) {
+            throw new IllegalArgumentException("O [name] não pode ser vazio, digite corretamente");
+        }
+
         repository.removeAccount(name);
     }
