+++ src/domain/usecase/entry/EntryExpenseUseCase.java
@@ -2,6 +2,7 @@ package domain.usecase.entry;
 
 import data.repository.entrys.EntryExpenseRepository;
 import domain.entities.entry.EntryExpenseEntities;
+import shared.Utils;
 
 import java.util.ArrayList;
 
@@ -12,8 +13,35 @@ public class EntryExpenseUseCase {
         repository = new EntryExpenseRepository();
     }
 
+    private void verifyAllIsEmpty(Boolean received, String receivedDate, Double valor, String date,
+                                  String category, String description) {
+
+        if (Utils.isEmptyParams(received.toString())) {
+            throw new IllegalArgumentException("O [received] não pode ser vazio, digite corretamente");
+        }
+        if (Utils.isEmptyParams(receivedDate)) {
+            throw new IllegalArgumentException("O [receivedDate] não pode ser vazio, digite corretamente");
+        }
+        if (Utils.isEmptyParams(valor.toString())) {
+            throw new IllegalArgumentException("O [valor] não pode ser vazio, digite corretamente");
+        }
+        if (Utils.isEmptyParams(date)) {
+            throw new IllegalArgumentException("O [date] não pode ser vazio, digite corretamente");
+        }
+        if (Utils.isEmptyParams(category)) {
+            throw new IllegalArgumentException("O [category] não pode ser vazio, digite corretamente");
+        }
+        if (Utils.isEmptyParams(description)) {
+            throw new IllegalArgumentException("O [description] não pode ser vazio, digite corretamente");
+        }
+
+    }
+
+
     public EntryExpenseEntities create(boolean received, String receivedDate, double valor, String date,
                                        String category, String description) {
+
+        verifyAllIsEmpty(received, receivedDate, valor, date, category, description);
 
         return repository.createEntryExpense(received, receivedDate, valor, date, category, description);
     }
@@ -21,10 +49,16 @@ public class EntryExpenseUseCase {
     public EntryExpenseEntities edit(boolean received, String receivedDate, String description, double valor,
                                      String date, String category) {
 
+        verifyAllIsEmpty(received, receivedDate, valor, date, category, description);
+
         return repository.editEntryExpense(received, receivedDate, description, valor, date, category);
     }
 
     public void remove(String description) {
+        if (Utils.isEmptyParams(description)) {
+            throw new IllegalArgumentException("O [description] não pode ser vazio, digite corretamente");
+        }
+
         repository.removeEntryExpense(description);
     }
